+++ site/docs/guides/changelog.md
@@ -6,8 +6,14 @@ shortTitle: Changelog
 
 ## VERSION 6.3.0
 
+#### `@wpackio/scripts`
+
 - **NEW** Option `useBabelConfig` in file config to override project-wide configuration for a particular entry.
 - **NEW** Option `compileNodeModules` in project config to handle how `node_modules` are compiled during production and development.
+
+#### `@wpackio/eslint-config`
+
+- **FIX** Babel plugin not being configured properly for typescript preset.
 
 ---